backend/requirements.txt
fastapi
uvicorn
moviepy
pydub
speechrecognition
google-cloud-speech
google-generativeai
rich
